Separate ProgressionInfo out from ForLoopLowering
